+++ purchases/src/test/java/com/revenuecat/purchases/paywalls/PaywallDataTest.kt
@@ -99,16 +99,6 @@ class PaywallDataTest {
         assertThat(localization.title).isEqualTo("Tienda")
     }
 
-    @Test
-    fun `paywall color can be created from a ColorInt`() {
-        val colorInt = Color.parseColor("#FFAABB")
-        val paywallColor = PaywallColor(colorInt)
-
-        assertThat(colorInt).isEqualTo(paywallColor.colorInt)
-        assertThat("#FFAABB").isEqualTo(paywallColor.stringRepresentation)
-        assertThat(Color.valueOf(colorInt)).isEqualTo(paywallColor.underlyingColor)
-    }
-
     private fun loadJSON(jsonFileName: String) = File(javaClass.classLoader!!.getResource(jsonFileName).file).readText()
 
 }
